move Header Bar to shared folder
